type system reboot part 1
This introduces a completely new TypeSystem to replace
what was previously spread across TypeSystem, Variables, and
RegsAndInstrs.

The new plan partly involves a new "NameId" which will become the
basis of name resolution. (essentially, replacing Ident with NameId
everywhere).  That unfortunately might involve writing some nasty
sort of "folding" function for a generic AST.  We'll see what happens.
Maybe NameId won't last and will be split up into more typesafe,
specific things like Func Ids and Var Ids, or something.

+++ src/ident.rs
@@ -27,7 +27,7 @@ impl Ident {
 }
 
 impl Ident {
-    pub fn as_ins(&self) -> Option<u32> {
+    pub fn as_ins(&self) -> Option<u16> {
         if let Some(remainder) = self.ident.strip_prefix("ins_") {
             Some(remainder.parse().expect("invalid instr ident, this is a bug!"))
         } else {
